authorization: Make verify_operation a method on AccessRuleList
