+++ crates/diom-authorization/tests/it/verification.rs
@@ -1,6 +1,4 @@
-use diom_authorization::{
-    AccessRuleList, Context, RequestedOperation, api::AccessRule, verify_operation,
-};
+use diom_authorization::{AccessRuleList, Context, RequestedOperation, api::AccessRule};
 use diom_id::Module;
 use serde_json::json;
 
@@ -58,7 +56,11 @@ fn test_verify_cache_example() {
         action: "get",
     };
     // no rule matches (need at least the extra /)
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_err());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_err()
+    );
 
     let op = RequestedOperation {
         module: Module::Cache,
@@ -67,7 +69,11 @@ fn test_verify_cache_example() {
         action: "get",
     };
     // simple match of first rule
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_ok());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_ok()
+    );
 
     let op = RequestedOperation {
         module: Module::Cache,
@@ -76,7 +82,11 @@ fn test_verify_cache_example() {
         action: "delete",
     };
     // no matching rule for action
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_err());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_err()
+    );
 
     let op = RequestedOperation {
         module: Module::Cache,
@@ -85,7 +95,11 @@ fn test_verify_cache_example() {
         action: "delete",
     };
     // wildcard action allowed for this key
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_ok());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_ok()
+    );
 
     let op = RequestedOperation {
         module: Module::Cache,
@@ -94,7 +108,11 @@ fn test_verify_cache_example() {
         action: "get",
     };
     // match of first rule with ** matching empty string
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_ok());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_ok()
+    );
 
     let op = RequestedOperation {
         module: Module::Cache,
@@ -103,7 +121,11 @@ fn test_verify_cache_example() {
         action: "get",
     };
     // explicit deny rule matches
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_err());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_err()
+    );
 
     let op = RequestedOperation {
         module: Module::Cache,
@@ -112,7 +134,11 @@ fn test_verify_cache_example() {
         action: "set",
     };
     // deny rule only affects Read, not List
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_ok());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_ok()
+    );
 
     let op = RequestedOperation {
         module: Module::Cache,
@@ -121,7 +147,11 @@ fn test_verify_cache_example() {
         action: "get",
     };
     // deny rule is exact-match, does not match here
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_ok());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_ok()
+    );
 
     let op = RequestedOperation {
         module: Module::Cache,
@@ -130,7 +160,11 @@ fn test_verify_cache_example() {
         action: "get",
     };
     // both cache rules only apply to the default namespace
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_err());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_err()
+    );
 
     let op = RequestedOperation {
         module: Module::Msgs,
@@ -139,7 +173,11 @@ fn test_verify_cache_example() {
         action: "get",
     };
     // no matching rule for module
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_err());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_err()
+    );
 }
 
 #[test]
@@ -153,7 +191,11 @@ fn test_verify_kv_example() {
         action: "delete",
     };
     // matches the first rule
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_ok());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_ok()
+    );
 
     let op = RequestedOperation {
         module: Module::Kv,
@@ -162,7 +204,11 @@ fn test_verify_kv_example() {
         action: "get",
     };
     // matches the second rule
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_ok());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_ok()
+    );
 
     let op = RequestedOperation {
         module: Module::Kv,
@@ -171,7 +217,11 @@ fn test_verify_kv_example() {
         action: "delete",
     };
     // Create action is only available for xyz namespace
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_err());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_err()
+    );
 
     let op = RequestedOperation {
         module: Module::Kv,
@@ -180,7 +230,11 @@ fn test_verify_kv_example() {
         action: "delete",
     };
     // same issue
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_err());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_err()
+    );
 
     let op = RequestedOperation {
         module: Module::Kv,
@@ -189,7 +243,11 @@ fn test_verify_kv_example() {
         action: "get",
     };
     // Read action does work for arbitrary namespaces, including the default one
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_ok());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_ok()
+    );
 
     let op = RequestedOperation {
         module: Module::Kv,
@@ -198,5 +256,9 @@ fn test_verify_kv_example() {
         action: "set",
     };
     // Same for List action, and a different named namespace
-    assert!(verify_operation(&op, &rules, Context::empty_for_tests()).is_ok());
+    assert!(
+        rules
+            .verify_operation(&op, Context::empty_for_tests())
+            .is_ok()
+    );
 }
