+++ crates/diom-proto/src/msgpack_or_json.rs
@@ -12,7 +12,7 @@ use axum::{
     response::{IntoResponse, Response},
 };
 use bytes::{BufMut as _, Bytes, BytesMut};
-use diom_authorization::{Context, Permissions, verify_operation};
+use diom_authorization::{Context, Permissions};
 use http::{HeaderMap, HeaderValue, StatusCode, header};
 use serde::{Serialize, de::DeserializeOwned};
 
@@ -153,7 +153,10 @@ where
 
         if let Some(op) = value.operation()
             && let Some(perms) = permissions
-            && verify_operation(&op, &perms.access_rules, Context::new(&perms)).is_err()
+            && perms
+                .access_rules
+                .verify_operation(&op, Context::new(&perms))
+                .is_err()
         {
             return Err(MsgPackOrJsonRejection::Forbidden {
                 resource: op.resource_str(),
